requirements.txt
pandas
Pillow
streamlit
scikit-learn